target/reporter/requirements.txt
# Core Dependencies
aiohttp>=3.8.0
pyyaml>=6.0
tiktoken>=0.4.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Financial Data
yfinance>=0.2.0

# Optional: For rich console output (used in some tools)
rich>=13.0.0